# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AfriChart EMR API",
        "description": "REST API for the AfriChart EMR — patients, consultations, prescriptions, invoices and the patient queue. All endpoints (except login) require a Sanctum bearer token.",
        "contact": {
            "name": "Lymora Tech",
            "email": "michael@mgbah.dev"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "List invoices (paginated)",
                "operationId": "176f86601dff5362250710476b314be6",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "issued",
                                "paid",
                                "partially_paid",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated invoice list"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/{invoice}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Show an invoice with line items",
                "operationId": "755978598358507deaa9cb9ad2f8ff4a",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice detail"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Update invoice tax / discount / notes",
                "operationId": "9eed7a7f666725292048e52baea4e1f2",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tax": {
                                        "type": "number"
                                    },
                                    "discount": {
                                        "type": "number"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/from-consultation/{consultation}": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Generate an invoice from a consultation",
                "operationId": "68ba4af1af260701ff45371e637a5536",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invoice generated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invoices/{invoice}/pay": {
            "patch": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Mark an invoice as paid",
                "operationId": "7ef8c800ca81714c0962c9ee88b591f6",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_method"
                                ],
                                "properties": {
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "transfer",
                                            "card",
                                            "insurance",
                                            "other"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marked paid"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/queue": {
            "get": {
                "tags": [
                    "Queue"
                ],
                "summary": "Today's patient queue",
                "operationId": "3be2515f12da45db4b74b74ade66bd52",
                "responses": {
                    "200": {
                        "description": "Queue entries"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Queue"
                ],
                "summary": "Check a patient into the queue",
                "operationId": "6b1ac87071aadac23b99ec76f77584b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "assigned_doctor_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "reason": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Checked in"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/queue/{queue}/assign": {
            "patch": {
                "tags": [
                    "Queue"
                ],
                "summary": "Assign a doctor to a queued patient",
                "operationId": "984d77bf6301ad7d146ee875752d2bce",
                "parameters": [
                    {
                        "name": "queue",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assigned_doctor_id"
                                ],
                                "properties": {
                                    "assigned_doctor_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doctor assigned"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/queue/{queue}/cancel": {
            "patch": {
                "tags": [
                    "Queue"
                ],
                "summary": "Cancel a queue entry",
                "operationId": "553c276a99447687831e511a88becc55",
                "parameters": [
                    {
                        "name": "queue",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancelled"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token returned by POST /auth/login.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}